+++ workon/errors.py
@@ -1,5 +1,0 @@
-"""Custom exceptions."""
-
-
-class ScriptError(Exception):
-    """Error in script."""